third_party/blink/web_tests/external/wpt/custom-elements/revamped-scoped-registry/CustomElementRegistry-upgrade.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] upgrade is a no-op when called on a shadow root with no association
  assert_equals: expected "HTMLElement" but got "GlobalABElement"
[FAIL] upgrade should upgrade a candidate element when called on a shadow root with an association
  registry.initialize is not a function
[FAIL] upgrade should not upgrade a candidate element not associated with the registry
  registry.initialize is not a function
Harness: the test ran to completion.
